Really fix polygon drawing
The previous fix (didn't make much sense on reflection and) was buggy
with the jumper animation.

This one does something sane, which is only include vertices in in the
"parity calculation" (determining when to draw) when it's the bottom
vertex of the line. This means that if two lines meet at the same
vertex:

1. If they're both bottoms then parity is ignored twice = no change
2. If they're both tops then parity is changed twice = no change
3. If one is a top and one is a bottom then parity is changed once =
start drawing if stopped and vice versa.

This means that local maxima and minima, and only local maxima and
minima, don't change the drawing calculation.