Add support for pluggable backends.
As committed, the backend fills in a struct for all platform-specific
operations, such as drawing, timing, getting input, etc.

This is actually a little too configurable: it implies that the binary
could be built with support for multiple backends. This is exactly what
you want for Linux, but for things like the Pebble watchface (and for
Mac OS and Windows, for that matter) it is overkill.