Add scene index and scene seeking
The scenes are used to generate a scene index consisting of (ms, offset)
tuples. These tuples are used when seeking to a particular millisecond
within the demo.

The seeking is still a little slow, since everything within the scene
must be replayed -- so for example if the scene contains multiple
animations, they must all be loaded. This could be improved but it may
not be worth it at the moment (scenes are small).